modify tools forms wizard
