+++ workflows/views/__init__.py
@@ -5,10 +5,9 @@ from django.shortcuts import render, redirect, get_object_or_404
 from formtools.wizard.views import SessionWizardView
 
 from galaxy.decorator import connection_galaxy
-from tools.forms import ToolForm
-from tools.models import Tool,ToolFlag
+from tools.models import ToolFlag
 from tools.views import tool_exec
-from workflows.models import Workflow
+from workflows.models import Workflow, WorkflowStepInformation
 from workspace.views import get_or_create_history
 
 
@@ -53,38 +52,6 @@ def launch_galaxy_workflow(request, slug):
     return redirect("galaxy_workflow")
 
 
-def form_class_list(gi, ids_tools):
-    """
-    Create ToolForm classes on the fly to be used by WizardForm
-    :param gi:
-    :param ids_tools:
-    :return: list af Class form
-    """
-    tools = Tool.objects.filter(pk__in=ids_tools)
-    tools_inputs_details = []
-    for tool in tools:
-        tool_inputs_details = gi.tools.show_tool(tool_id=tool.id_galaxy, io_details='true')
-        tools_inputs_details.append(
-            type(str(tool.name) + 'Form',
-                 (ToolForm,),
-                 {'tool_params': tool_inputs_details.get('inputs'),'tool_id': tool.id_galaxy }
-                 )
-        )
-    return tools_inputs_details
-
-
-@connection_galaxy
-def workflow_form(request):
-
-    tools = Tool.objects.all()
-    form_list = form_class_list(request.galaxy, tools)
-    selected_tools = request.session.get('selected_tools')
-
-    if selected_tools:
-         form_list = form_class_list(request.galaxy, selected_tools)
-
-    return WorkflowWizard.as_view(form_list=form_list)(request)
-
 
 class WorkflowWizard(SessionWizardView):
 
@@ -97,6 +64,38 @@ class WorkflowWizard(SessionWizardView):
             print output
 
         return redirect(reverse_lazy("history_detail", kwargs={'history_id': output }))
+
+
+def form_class_list(galaxy_server, tools):
+    """
+    Create ToolForm classes on the fly to be used by WizardForm
+    :param gi:
+    :param tools:
+    :return: list af Class form
+    """
+    tools_inputs_details = []
+    for tool in tools:
+        tools_inputs_details.append(tool.form_class(galaxy_server=galaxy_server))
+    return tools_inputs_details
+
+
+@connection_galaxy
+def workflow_form(request, slug_workflow):
+
+    gi = request.galaxy
+    workflow= Workflow.objects.get(slug=slug_workflow)
+    workflow_json = gi.workflows.show_workflow(workflow_id=workflow.id_galaxy)
+    tools = [ t[1] for t in WorkflowStepInformation(workflow_json).sorted_tool_list]
+    # parse galaxy workflows json informations
+    form_list = form_class_list(request.galaxy_server, tools )
+    selected_tools = request.session.get('selected_tools')
+
+    if selected_tools:
+         form_list = form_class_list(request.galaxy_server, selected_tools)
+
+    ClassWizardView = type(str(slug_workflow)+"Wizard", (WorkflowWizard,), {'form_list': form_list} )
+
+    return ClassWizardView.as_view()(request)
